chore: move server handler to routes/api/v1
